requirements.txt
fastapi==0.115.2
fastapi-utils==0.7.0
SQLAlchemy==2.0.36
uvicorn==0.32.0
python-jose==3.3.0
typing_inspect==0.9.0
PyJWT==2.9.0
python-multipart==0.0.12
